Fix pluralization of "walrus"
The word "walrii", meant to be a plural form of "walrus", sparks some
interesting discussions of how that word may have come to be:

Using an "i" to pluralize a noun that ends in "us" employs the rule that
applies to many Latin masculine nouns of the second declension use.
But for that to yield "walrii" the singular form would have to be
"walrius", which conflices with the singular form "walrus" used in the
for loop on the very next line.

"Virii" gets away with this because it isn't clear how to properly
pluralize a second declension neuter mass noun.  "Walrus" isn't a mass
noun, though, so it doesn't get to use that trick.

Finally, "walrus" is Dutch.  We shouldn't be pluralizing it like a
Latin word.  8^)